updated testcases wrt gunicorn changes and refracted the code
