+++ mlflow/pyfunc/constants.py
@@ -1,0 +1,13 @@
+CONTENT_TYPE_CSV = "text/csv"
+CONTENT_TYPE_JSON = "application/json"
+CONTENT_TYPE_JSON_RECORDS_ORIENTED = "application/json; format=pandas-records"
+CONTENT_TYPE_JSON_SPLIT_ORIENTED = "application/json; format=pandas-split"
+
+MODEL_ARTIFACT_PATH_VAR = 'MLFLOW_MODEL_ARTIFACT_PATH'
+
+CONTENT_TYPES = [
+    CONTENT_TYPE_CSV,
+    CONTENT_TYPE_JSON,
+    CONTENT_TYPE_JSON_RECORDS_ORIENTED,
+    CONTENT_TYPE_JSON_SPLIT_ORIENTED
+]
